+++ supertokens/constants.go
@@ -21,8 +21,8 @@ const (
 )
 
 // VERSION current version of the lib
-const VERSION = "0.8.2"
+const VERSION = "0.8.3"
 
 var (
-	cdiSupported = []string{"2.8", "2.9", "2.10", "2.11", "2.12", "2.13", "2.14"}
+	cdiSupported = []string{"2.8", "2.9", "2.10", "2.11", "2.12", "2.13", "2.14", "2.15"}
 )